racoonsmeal-backend/pyproject.toml
[project]
name = "racoonsmeal-backend"
version = "0.1.0"
description = "A meal planner that gives you the most random meals for your day but balances calories & nutrients for the day"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-storages[s3]>=1.14.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
prod = [
    "gunicorn>=23.0.0",
]
